Add files via upload
the code above is a python program that ask a user to input his/age, if the users age is less than 18
the program will return "you are a minor". else if the user enters an age that is greater than 18 and less
than 65, it will return "Your are an adult". else if the user enters an age greater than 65 years it will 
output "you are a senior citizen".